# docs/swagger.yaml
basePath: /api
definitions:
  github_com_amalgamated-tools_biblioteka_internal_db.Timestamp:
    properties:
      time.Time:
        type: string
    type: object
...
  internal_handlers.authorDTO:
    properties:
      created_at:
        $ref: '#/definitions/github_com_amalgamated-tools_biblioteka_internal_db.Timestamp'
      goodreads_id:
        type: string
      google_books_id:
        type: string
      hardcover_id:
        type: string
      id:
        type: string
      image_url:
        type: string
      name:
        type: string
      updated_at:
        $ref: '#/definitions/github_com_amalgamated-tools_biblioteka_internal_db.Timestamp'
    type: object
  internal_handlers.authorRequest:
    properties:
      goodreads_id:
        type: string
      google_books_id:
        type: string
      hardcover_id:
        type: string
      image_url:
        type: string
      name:
        type: string
    type: object
...
  internal_handlers.bookFileDTO:
    properties:
      book_id:
        type: string
      created_at:
        $ref: '#/definitions/github_com_amalgamated-tools_biblioteka_internal_db.Timestamp'
      file_hash:
        type: string
      file_name:
        type: string
      file_path:
        type: string
      file_size:
        type: integer
      file_type:
        type: string
      id:
        type: string
      updated_at:
        $ref: '#/definitions/github_com_amalgamated-tools_biblioteka_internal_db.Timestamp'
    type: object
...
  internal_handlers.errorResponse:
    properties:
      error:
        example: error message
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Personal library management API
  title: Biblioteka API
  version: "1.0"
paths:
...
  /authors:
    get:
      description: Returns all authors
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_handlers.authorDTO'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
      security:
      - BearerAuth: []
      summary: List authors
      tags:
      - Authors
    post:
      consumes:
      - application/json
      description: Create a new author
      parameters:
      - description: Author data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.authorRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_handlers.authorDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
      security:
      - BearerAuth: []
      summary: Create an author
      tags:
      - Authors
  /authors/{id}:
    delete:
      description: Delete an author by ID
      parameters:
      - description: Author ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
      security:
      - BearerAuth: []
      summary: Delete an author
      tags:
      - Authors
    get:
      description: Returns a single author by ID
      parameters:
      - description: Author ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.authorDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
      security:
      - BearerAuth: []
      summary: Get an author
      tags:
      - Authors
    put:
      consumes:
      - application/json
      description: Update an existing author
      parameters:
      - description: Author ID
        in: path
        name: id
        required: true
        type: string
      - description: Author data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.authorRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.authorDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
      security:
      - BearerAuth: []
      summary: Update an author
      tags:
      - Authors
  /book-files/{id}:
    delete:
      description: Delete a book file by ID
      parameters:
      - description: Book File ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
      security:
      - BearerAuth: []
      summary: Delete a book file
      tags:
      - BookFiles
    get:
      description: Returns a single book file by ID
      parameters:
      - description: Book File ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.bookFileDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
      security:
      - BearerAuth: []
      summary: Get a book file
      tags:
      - BookFiles
...
securityDefinitions:
  BearerAuth:
    description: Enter "Bearer {token}"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"